Merge branch 'removechat'

+++ graphic/geometry.h
@@ -8,7 +8,7 @@ public:
 	static unsigned WindowWidth;
 	static unsigned WindowHeight;
 	static float WindowScale() {return static_cast<float>(WindowHeight) / static_cast<float>(BaseSize);}
-	static unsigned SidebarWidth() {return static_cast<unsigned>((WindowWidth - WindowHeight) / WindowScale());}
+	static unsigned SidebarWidth() {return static_cast<unsigned>(static_cast<float>(WindowWidth - WindowHeight) / WindowScale());}
 public: /* Monostate class: cannot be instantiated */
 	Geometry() = delete;
 	Geometry(const Geometry&) = delete;
